// src/endpoints/common-types.ts
export const weakUUIDSchema = z
  .string()
  .regex(/^([a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12})$/i)
  .describe("UUID");
// ...
export const dateSchema = z
  .string()
  .datetime()
  .transform((val) => new Date(val))
  .describe("Date in ISO 8601 format");
// ...
export const itemIDSchema = weakUUIDSchema.describe("Item ID");
export const itemTypeIDSchema = weakUUIDSchema.describe("Item Type ID");
export const armorIDSchema = weakUUIDSchema.describe("Armor ID");
export const currencyIDSchema = weakUUIDSchema.describe("Currency ID");
// ...
export const offerSchema = z.object({
  OfferID: z.string(),
  IsDirectPurchase: z.boolean(),
  StartDate: dateSchema,
  Cost: z.record(currencyIDSchema, z.number()),
  Rewards: z.array(
    z.object({
      ItemTypeID: itemTypeIDSchema,
      ItemID: itemIDSchema,
      Quantity: z.number(),
    }),
  ),
});